+++ source/src/supermarket/model/customer/Podge.java
@@ -1,7 +1,7 @@
 package supermarket.model.customer;
 
 public class Podge extends Customer{
-	public Podge(int spawnX, int spawnY){
+	public Podge(double spawnX, double spawnY){
 		super(spawnX, spawnY);
 	}
 	@Override
