+++ source/src/supermarket/model/customer/Student.java
@@ -1,7 +1,7 @@
 package supermarket.model.customer;
 
 public class Student extends Customer{
-	public Student(int spawnX, int spawnY){
+	public Student(double spawnX, double spawnY){
 		super(spawnX, spawnY);
 	}
 	@Override
